+++ lib/object_inspector/formatters/templating_formatter.rb
@@ -41,10 +41,18 @@ module ObjectInspector
     #
     # @return [String]
     def call
-      build_string
+      if wrapped_object_inspection
+        build_wrapped_object_string
+      else
+        build_string
+      end
     end
 
   private
+
+    def build_wrapped_object_string
+      "#{build_string} #{RIGHT_ARROW_ICON} #{wrapped_object_inspection}"
+    end
 
     def build_string
       if flags